refactor(State): Finished가 Ready를 상속받도록 수정
- 예외 메시지 추가
