+++ src/main/java/chess/domain/state/BlackWin.java
@@ -1,4 +1,10 @@
 package chess.domain.state;
 
+import chess.domain.board.Board;
+
 public class BlackWin extends Finished {
+
+    BlackWin(Board board) {
+        super(board);
+    }
 }
