+++ src/main/java/chess/domain/state/Ready.java
@@ -1,6 +1,11 @@
 package chess.domain.state;
 
 import chess.domain.board.Board;
+import chess.domain.board.Score;
+import chess.domain.board.ScoreResult;
+import chess.domain.piece.Color;
+import java.util.HashMap;
+import java.util.Map;
 
 public abstract class Ready implements State {
 
@@ -17,5 +22,17 @@ public abstract class Ready implements State {
     @Override
     public final Board getBoard() {
         return new Board(board.getValue());
+    }
+
+    @Override
+    public final ScoreResult getScore() {
+        double whiteScore = board.calculateScore(Color.WHITE);
+        double blackScore = board.calculateScore(Color.BLACK);
+
+        Map<Color, Score> scoreByColor = new HashMap<>();
+        scoreByColor.put(Color.WHITE, new Score(whiteScore));
+        scoreByColor.put(Color.BLACK, new Score(blackScore));
+
+        return new ScoreResult(scoreByColor);
     }
 }
