+++ src/main/java/chess/domain/state/WhiteTurn.java
@@ -14,7 +14,7 @@ final class WhiteTurn extends Running {
     public State movePiece(Position src, Position dest) {
         board.move(src, dest, Color.WHITE);
         if (!board.hasKing(Color.BLACK)) {
-            return new WhiteWin();
+            return new WhiteWin(board);
         }
         return new BlackTurn(board);
     }
