+++ src/main/java/chess/domain/state/WhiteWin.java
@@ -1,4 +1,10 @@
 package chess.domain.state;
 
+import chess.domain.board.Board;
+
 public class WhiteWin extends Finished {
+
+    WhiteWin(Board board) {
+        super(board);
+    }
 }
